refactor: Cars 모델 생성, 컨트롤러 함수 분리
